+++ src/SmartFormat.Tests/Extensions/ReflectionSourceTests.cs
@@ -78,9 +78,9 @@ namespace SmartFormat.Tests.Extensions
             var format = "{0} {0.ToLower} {ToLower} {ToUpper}";
             //var expected = "Zero zero zero ZERO";
 
-            var smart = new SmartFormatter();
-            smart.AddExtensions(new ReflectionSource(), new DefaultSource());
-            smart.AddExtensions(new DefaultFormatter());
+            var smart = new SmartFormatter()
+                .AddExtensions(new ReflectionSource(), new DefaultSource())
+                .AddExtensions(new DefaultFormatter());
 
             var args = GetArgs();
             Assert.That(() => smart.Format(format, args), Throws.Exception.TypeOf(typeof(FormattingException)).And.Message.Contains("ToLower"));
@@ -92,9 +92,9 @@ namespace SmartFormat.Tests.Extensions
         [Test]
         public void Test_Methods_CaseInsensitive()
         {
-            var smart = new SmartFormatter(new SmartSettings{ CaseSensitivity = CaseSensitivityType.CaseInsensitive });
-            smart.AddExtensions(new ReflectionSource());
-            smart.AddExtensions(new DefaultFormatter());
+            var smart = new SmartFormatter(new SmartSettings { CaseSensitivity = CaseSensitivityType.CaseInsensitive })
+                .AddExtensions(new ReflectionSource())
+                .AddExtensions(new DefaultFormatter());
 
             var format = "{0} {0.ToLower} {toloWer} {touPPer}";
             //var expected = "Zero zero zero ZERO";
@@ -105,27 +105,27 @@ namespace SmartFormat.Tests.Extensions
         [Test]
         public void Void_Methods_Should_Just_Be_Ignored()
         {
-            var smart = new SmartFormatter();
-            smart.AddExtensions(new ReflectionSource(), new DefaultSource());
-            smart.AddExtensions(new DefaultFormatter());
+            var smart = new SmartFormatter()
+                .AddExtensions(new ReflectionSource(), new DefaultSource())
+                .AddExtensions(new DefaultFormatter());
             Assert.That(() => smart.Format("{0.Clear}", smart.SourceExtensions), Throws.Exception.TypeOf(typeof(FormattingException)).And.Message.Contains("Clear"));
         }
 
         [Test]
         public void Methods_With_Parameter_Should_Just_Be_Ignored()
         {
-            var smart = new SmartFormatter();
-            smart.AddExtensions(new ReflectionSource(), new DefaultSource());
-            smart.AddExtensions(new DefaultFormatter());
+            var smart = new SmartFormatter()
+                .AddExtensions(new ReflectionSource(), new DefaultSource())
+                .AddExtensions(new DefaultFormatter());
             Assert.That(() => smart.Format("{0.Add}", smart.SourceExtensions), Throws.Exception.TypeOf(typeof(FormattingException)).And.Message.Contains("Add"));
         }
 
         [Test]
         public void Properties_With_No_Getter_Should_Just_Be_Ignored()
         {
-            var smart = new SmartFormatter();
-            smart.AddExtensions(new ReflectionSource(), new DefaultSource());
-            smart.AddExtensions(new DefaultFormatter());
+            var smart = new SmartFormatter()
+                .AddExtensions(new ReflectionSource(), new DefaultSource())
+                .AddExtensions(new DefaultFormatter());
             Assert.That(() => smart.Format("{Misc.OnlySetterProperty}", new { Misc = new MiscObject() }), Throws.Exception.TypeOf(typeof(FormattingException)).And.Message.Contains("OnlySetterProperty"));
         }