+++ src/SmartFormat/Smart.cs
@@ -96,56 +96,60 @@ namespace SmartFormat
         /// <summary>
         /// Gets or sets the default <see cref="SmartFormatter"/>.
         /// If not set, the <see cref="CreateDefaultSmartFormat"/> will be used.
-        /// It is recommended to set the <see cref="Default"/> <see cref="SmartFormatter"/> with the extensions that are actually needed.
+        /// It is recommended to set the <see langword="static"/> <see cref="Default"/> <see cref="SmartFormatter"/> with the extensions that are actually needed.
         /// </summary>
         public static SmartFormatter Default { get; set; } = CreateDefaultSmartFormat();
 
         /// <summary>
-        /// Creates a <see cref="SmartFormatter"/> with all extensions registered.
+        /// Creates a new <see cref="SmartFormatter"/> instance with core extensions registered.
         /// For optimized performance, create a <see cref="SmartFormatter"/> instance and register the
-        /// particular extensions that are needed.
+        /// particular extensions that are really needed.
+        /// <para>
+        /// See <see cref="WellKnownExtensionTypes.Formatters"/> and <see cref="WellKnownExtensionTypes.Sources"/>
+        /// for a complete list of well-known types.
+        /// </para>
         /// </summary>
         /// <param name="settings">The <see cref="SmartSettings"/> to use, or <see langword="null"/> for default settings.</param>
-        /// <returns>A <see cref="SmartFormatter"/> with all extensions registered.</returns>
+        /// <returns>A <see cref="SmartFormatter"/> with core extensions registered:
+        /// <para>
+        /// <see cref="ISource"/>s:
+        /// <see cref="StringSource"/>, <see cref="ListFormatter"/>, <see cref="DictionarySource"/>,
+        /// <see cref="ValueTupleSource"/>, <see cref="ReflectionSource"/>, <see cref="DefaultSource"/>.
+        /// </para>
+        /// <para>
+        /// <see cref="IFormatter"/>s:
+        /// <see cref="ListFormatter"/>, <see cref="PluralLocalizationFormatter"/>,
+        /// <see cref="ConditionalFormatter"/>, <see cref="IsMatchFormatter"/>, <see cref="NullFormatter"/>,
+        /// <see cref="ChooseFormatter"/>, <see cref="SubStringFormatter"/>, <see cref="DefaultFormatter"/>.
+        /// </para>
+        /// </returns>
         public static SmartFormatter CreateDefaultSmartFormat(SmartSettings? settings = null)
         {
             // Register all default extensions here:
-            var formatter = new SmartFormatter(settings);
-            
-            // Add all extensions:
-            // Note, the order is important; the extensions
-            // will be executed in this order:
-
-            var listSourceAndFormatter = new ListFormatter();
-
-            // sources for specific types must be in the list before ReflectionSource
-            formatter.AddExtensions(
+            var smart = new SmartFormatter(settings)
+            // Extension are sorted automatically
+            .AddExtensions(
                 new StringSource(),
-                (ISource) listSourceAndFormatter, // ListFormatter should be one of the first source extensions
+                // will automatically be added to the IFormatter list, too
+                new ListFormatter(),
                 new DictionarySource(),
                 new ValueTupleSource(),
-                new SystemTextJsonSource(),
-                new NewtonsoftJsonSource(),
-                new XmlSource(),
                 new ReflectionSource(),
-
-                // The DefaultSource reproduces the string.Format behavior:
+                // for string.Format behavior
                 new DefaultSource()
-            );
-            formatter.AddExtensions(
-                (IFormatter) listSourceAndFormatter, // ListFormatter should be one of the first formatter extensions
+            )
+            .AddExtensions(
                 new PluralLocalizationFormatter(),
                 new ConditionalFormatter(),
                 new IsMatchFormatter(),
                 new NullFormatter(),
-                new TimeFormatter(),
-                new XElementFormatter(),
                 new ChooseFormatter(),
                 new SubStringFormatter(),
+                // for string.Format behavior
                 new DefaultFormatter()
             );
 
-            return formatter;
+            return smart;
         }
 
         #endregion